src/server/requirements.txt
Flask
Flask-PyMongo
flasgger
python-dotenv